feat(FN-2535): move get formatted report period helper to common
