--- libs/common/src/helpers/date.ts
@@ -1,6 +1,0 @@
-/**
- * Gets the one indexed month
- * @param dateInMonth - A date in the month
- * @returns The one indexed month
- */
-export const getOneIndexedMonth = (dateInMonth: Date) => dateInMonth.getMonth() + 1;
