+++ portal/server/helpers/index.js
@@ -18,8 +18,7 @@ const validationErrorHandler = require('./validationErrorHandler');
 const constructPayload = require('./constructPayload');
 const { getOneIndexedMonth } = require('./getOneIndexedMonth');
 const { getBusinessDayOfMonth } = require('./getBusinessDayOfMonth');
-const { getNowAsEpoch, getDateFromMonthAndYear } = require('./date');
-const { getFormattedReportPeriod } = require('./getFormattedReportPeriod');
+const { getNowAsEpoch } = require('./date');
 
 module.exports = {
   isEveryDealFormComplete,
@@ -44,6 +43,4 @@ module.exports = {
   getOneIndexedMonth,
   getBusinessDayOfMonth,
   getNowAsEpoch,
-  getDateFromMonthAndYear,
-  getFormattedReportPeriod,
 };
